pyproject.toml
[project]
name = "doc-manager-mcp"
version = "0.1.0"
description = "MCP server for comprehensive documentation lifecycle management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
doc-manager = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
